requirements.txt

email-validator>=2.3.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
pillow>=11.3.0
psycopg2-binary>=2.9.10
werkzeug>=3.1.3
